Se crea el método de búsqueda por Query y el Update, falta probar

+++ Estructuras/Automovil.cs
@@ -9,7 +9,7 @@ namespace Estructuras
         public ObjectId? ID { get; set; }
 
         [JsonPropertyName("idAuto")]
-        public long IdAuto { get; set; }
+        public int IdAuto { get; set; }
 
         [JsonPropertyName("marca")]
         public string? Marca { get; set; }
@@ -18,7 +18,7 @@ namespace Estructuras
         public string? Modelo { get; set; }
 
         [JsonPropertyName("año")]
-        public long Año { get; set; }
+        public int Año { get; set; }
 
         [JsonPropertyName("serie")]
         public string? Serie { get; set; }
@@ -28,7 +28,7 @@ namespace Estructuras
         }
 
         [BsonCtor]
-        public Automovil(ObjectId _id, long idAuto, string marca, string modelo, long año, string serie)
+        public Automovil(ObjectId _id, int idAuto, string marca, string modelo, int año, string serie)
         {
             ID = _id;
             IdAuto = idAuto;
